# Conflicts:
#	src/main/java/ru/innopolis/university/summerbootcamp2016/poker/Game.java

+++ src/main/java/ru/innopolis/university/summerbootcamp2016/poker/Bank.java
@@ -11,7 +11,7 @@ public class Bank {
         this.bankBalance += bankBalance;
     }
     public long getReward(){
-        int bal = this.bankBalance;
+        long bal = this.bankBalance;
         this.bankBalance = 0;
         return bal;
     }
